Problem 008: Largest product in a series.
Python implementation is proof of concept. Simpler but less efficient,
as it slices the input and computes product each time.
C implementation is a sliding window that divides the working product by
the first digit and multiplies by the last, so that the whole product
need not be recomputed each time.